+++ src/components/Features2.jsx
@@ -18,15 +18,15 @@ export const Features2 = () => (
             <div className="mb-4 py-3 md:pl-3 md:pr-20 lg:pr-12 rounded">
               <img
                 src={feature5.src}
-                alt="f1"
-                className="rounded-xl  custom-border-gray"
+                alt="Feature image 5"
+                className="rounded-xl  main-border-gray"
               />
             </div>
             <div className="py-3 md:pl-20 lg:pl-12 md:pr-2 rounded ">
               <img
                 src={feature6.src}
-                alt="f2"
-                className="rounded-xl  custom-border-gray"
+                alt="Feature image 6"
+                className="rounded-xl  main-border-gray"
               />
             </div>
           </div>
@@ -34,18 +34,16 @@ export const Features2 = () => (
 
         <div className="w-full lg:w-1/2 mb-12 lg:mb-0 xl:pl-8">
           <div className="mx-auto lg:mx-auto w-11/12 sm:w-4/5 md:w-3/4 lg:w-unset">
-            <span className="block-subtitle">
-              Make Data-Driven Decisions
-            </span>
+            <span className="block-subtitle">Make Data-Driven Decisions</span>
             <h2 className="mt-6 mb-8 text-4xl lg:text-5xl block-big-title">
               Issue tracking you’ll enjoy using
             </h2>
-            <p className="mb-12 text-customGrayText leading-loose">
+            <p className="mb-12 text-secondaryText leading-loose">
               Monitor and track data issues with ease using our intuitive and
               efficient issue tracking system. Stay ahead of potential problems
               and improve your workflow.
             </p>
-            <ul className="mb-6 text-white">
+            <ul className="mb-6 text-primaryText">
               <li className="mb-4 flex">
                 <CheckArrowIcon />
                 <span>Collaborative environment</span>