+++ src/components/FeaturesDiagonal.jsx
@@ -9,17 +9,17 @@ export const FeaturesDiagonal = () => {
 
   return (
     <section className="lg:mb-16 w-full flex flex-col justify-center items-center bg-bgDark1">
-      <div className="custom-shape-divider-bottom-1665696614">
+      <div className="shape-divider-bottom-1665696614">
         <svg
           data-name="Layer 1"
           xmlns="http://www.w3.org/2000/svg"
           viewBox="0 0 1200 120"
           preserveAspectRatio="none"
-          className="custom-bg-dark2"
+          className="bg-bgDark2  fill-bgDark2"
         >
           <path
             d="M1200 120L0 16.48 0 0 1200 0 1200 120z"
-            className="custom-bg-dark1"
+            className="bg-bgDark1  fill-bgDark1"
           ></path>
         </svg>
       </div>
@@ -31,44 +31,43 @@ export const FeaturesDiagonal = () => {
       >
         <div className=" 2xl:w-[1150px] xl:w-[1050px]  md:w-4/5 flex justify-center bg-bgDark1 pt-12 lg:pt-24 pb-8 lg:pb-20 mx-auto lg:flex-row flex-col">
           <div className="w-3/4 lg:w-1/2 flex flex-col lg:mx-unset mx-auto">
-            <span className="block-subtitle">
-              Accelerate Your Success
-            </span>
+            <span className="block-subtitle">Accelerate Your Success</span>
             <h2 className="mt-10 mb-8 text-4xl lg:text-5xl block-big-title">
               Build &amp; Launch without problems
             </h2>
-            <p className="mb-16 text-customGrayText leading-loose">
+            <p className="mb-16 text-secondaryText leading-loose">
               Our platform enables you to launch your data-driven projects with
               ease. Boost productivity and achieve better results. Empower your
               decision-making with advanced analytics
             </p>
-            <div
+            <button
               className="w-[210px] h-12 contained-button mr-10 "
               onClick={() => setIsModalOpen(true)}
+              aria-label="Get started"
             >
               Get Started
-            </div>
+            </button>
           </div>
           <div className="w-4/5 lg:w-1/2 lg:pl-16 flex justify-center mx-auto pt-16 lg:pt-0">
             <img
               src={featuresdiagonal.src}
-              alt="f1"
-              className="rounded-xl  custom-border-gray"
+              alt="Feature image"
+              className="rounded-xl  main-border-gray"
             />
           </div>
         </div>
       </motion.div>
-      <div className="custom-shape-divider-top-1665696661 w-full">
+      <div className="shape-divider-top-1665696661 w-full">
         <svg
           data-name="Layer 1"
           xmlns="http://www.w3.org/2000/svg"
           viewBox="0 0 1200 120"
           preserveAspectRatio="none"
-          className="custom-bg-dark2"
+          className="bg-bgDark2 fill-bgDark2"
         >
           <path
             d="M1200 120L0 16.48 0 0 1200 0 1200 120z"
-            className="custom-bg-dark1"
+            className="bg-bgDark1 fill-bgDark1"
           ></path>
         </svg>
       </div>