+++ src/components/Footer.jsx
@@ -27,7 +27,7 @@ const footerData = [
 
 export const Footer = () => {
   return (
-    <footer>
+    <footer aria-label="Site footer">
       <div className="pt-10  lg:pt-20 lg:pb-16 bg-bgDark1 radius-for-skewed ">
         <div className="container mx-auto px-4 w-4/5 md:w-11/12 lg:w-10/12 xl:w-4/5 2xl:w-2/3">
           <div className="flex flex-wrap">
@@ -47,27 +47,30 @@ export const Footer = () => {
                   href="https://bootstrapious.com/p/bootstrap-4-dark-admin"
                   target="_blank"
                   className="text-gray-100 ml-1.5 "
-                  aria-label="DarkAdmin"
+                  aria-label="Dark Admin Dashboard"
                 >
                   here.
                 </a>
               </p>
               <div className="w-36 mx-auto lg:mx-0">
                 <a
-                  className="inline-block w-10  h-10 mr-2 p-2 bg-bgDark2 custom-border-gray  hover:bg-gray-700 rounded-xl"
+                  className="inline-block w-10  h-10 mr-2 p-2 pt-[0.55rem] outlined-button"
                   href="#"
+                  aria-label="Facebook"
                 >
                   <FacebookIcon />
                 </a>
                 <a
-                  className="inline-block w-10  h-10 mr-2 p-2 bg-bgDark2 custom-border-gray  hover:bg-gray-700 rounded-xl"
+                  className="inline-block w-10  h-10 mr-2 p-2 pt-[0.55rem] pl-[0.55rem] outlined-button"
                   href="#"
+                  aria-label="Twitter"
                 >
                   <TwitterIcon />
                 </a>
                 <a
-                  className="inline-block w-10  h-10 mr-2 p-2 bg-bgDark2 custom-border-gray  hover:bg-gray-700 rounded-xl"
+                  className="inline-block w-10  h-10 mr-2 p-2 pt-[0.55rem] pl-[0.55rem] outlined-button"
                   href="#"
+                  aria-label="Instagram"
                 >
                   <InstagramIcon />
                 </a>
@@ -75,14 +78,14 @@ export const Footer = () => {
             </div>
             <div className="w-full lg:w-2/3  lg:pl-16 hidden lg:flex flex-wrap justify-between">
               <div className="w-full md:w-1/3 lg:w-auto mb-16 md:mb-0">
-                <h3 className="mb-6 text-2xl font-bold text-white">Products</h3>
+                <h3 className="mb-6 text-2xl font-bold text-primaryText">Products</h3>
                 <ul>
                   {footerData[0].items.map((item, index) => (
                     <li key={`${item}-${index}`} className="mb-4">
                       <a
                         className="text-gray-400 hover:text-gray-300"
                         href="#"
-                        aria-label=""
+                        aria-label={item}
                       >
                         {item}
                       </a>
@@ -91,7 +94,7 @@ export const Footer = () => {
                 </ul>
               </div>
               <div className="w-full md:w-1/3 lg:w-auto mb-16 md:mb-0">
-                <h3 className="mb-6 text-2xl font-bold text-white">
+                <h3 className="mb-6 text-2xl font-bold text-primaryText">
                   Important Links
                 </h3>
                 <ul>
@@ -100,7 +103,7 @@ export const Footer = () => {
                       <a
                         className="text-gray-400 hover:text-gray-300"
                         href="#"
-                        aria-label=""
+                        aria-label={item}
                       >
                         {item}
                       </a>
@@ -109,14 +112,14 @@ export const Footer = () => {
                 </ul>
               </div>
               <div className="w-full md:w-1/3 lg:w-auto">
-                <h3 className="mb-6 text-2xl font-bold text-white">Company</h3>
+                <h3 className="mb-6 text-2xl font-bold text-primaryText">Company</h3>
                 <ul>
                   {footerData[2].items.map((item, index) => (
                     <li key={`${item}-${index}`} className="mb-4">
                       <a
                         className="text-gray-400 hover:text-gray-300"
                         href="#"
-                        aria-label=""
+                        aria-label={item}
                       >
                         {item}
                       </a>
@@ -125,7 +128,7 @@ export const Footer = () => {
                 </ul>
               </div>
             </div>
-          </div>          
+          </div>
         </div>
       </div>
     </footer>