+++ src/components/Testimonials.jsx
@@ -39,9 +39,7 @@ export const Testimonials = () => (
         viewport={{ once: true }}
         transition={{ duration: 0.4, delay: 0.3 }}
       >
-        <div className="block-subtitle text-center mb-6">
-          Testimonials
-        </div>
+        <div className="block-subtitle text-center mb-6">Testimonials</div>
         <div className="block-big-title text-center mb-20 px-8 sm:px-24 md:px-48">
           People like you love Tailcast
         </div>
@@ -49,17 +47,23 @@ export const Testimonials = () => (
         <div className="flex flex-col lg:flex-row gap-8 lg:gap-5 xl:gap-10 px-6 xl:px-0 items-center">
           {testimonialsData.map((testimonial, index) => (
             <div
-              className="w-11/12 sm:w-4/5 md:w-[560px] lg:w-1/3 custom-border-gray-darker rounded-xl bg-bgDark3 flex flex-col px-6 py-4"
+              className="w-11/12 sm:w-4/5 md:w-[560px] lg:w-1/3 main-border-gray-darker rounded-xl bg-bgDark3 flex flex-col px-6 py-4"
               key={`${testimonial.customerName}-${index}`}
             >
               <div className="flex mb-2">
                 <QuoteIcon />
               </div>
-              <div className="content-text-white">
-                "{testimonial.content}"
-              </div>
+              <div className="content-text-white">"{testimonial.content}"</div>
               <div className="flex mt-4 mb-2 xl:mt-8 xl:mb-4">
-                <img src={testimonial.image.src} alt="" width="45px" />
+                <div>
+                  <img
+                    src={testimonial.image.src}
+                    alt="Customer avatar"
+                    width="45px"
+                    height="5px"
+                    aria-label={testimonial.customerName}
+                  />
+                </div>
                 <div className="flex flex-col ml-4">
                   <div className="content-text-white font-medium">
                     {testimonial.customerName}