+++ tailwind.config.cjs
@@ -6,15 +6,18 @@ module.exports = {
       colors: {
         primaryColor: "rgb(99, 102, 241)",
         secondaryColor: "rgb(161, 163, 247)",
+        primaryText: "rgb(255,255,255)",
+        secondaryText: "rgb(174, 178, 183)",
         bgDark1: "rgb(31, 32, 35)",
         bgDark2: "rgb(38, 39, 43)",
         bgDark3: "rgb(48, 49, 54)",
         bgDark3Hover: "rgb(55, 56, 62)",
-        quoteIconColor: "rgb(178, 184, 205)",
-        customGrayText: "rgb(174, 178, 183)",
         bgDarkTransparent: "rgb(31, 32, 35, 0.7)",
         bgDarkTransparentDarker: "rgb(0,0,0,0.5)",
-        bgDarkTransparentLighter: "rgb(48, 49, 54, 0.7)"  
+        bgDarkTransparentLighter: "rgb(48, 49, 54, 0.7)",
+        mainBorder: "rgb(255,255,255,0.15)",
+        mainBorderDarker: "rgb(255,255,255,0.07)",
+        quoteIconColor: "rgb(178, 184, 205)"
       },
       fontFamily: {
         Inter: "Inter",
@@ -24,8 +27,7 @@ module.exports = {
         sm: "640px",
         md: "768px",
         lg: "1024px",
-        xl: "1280px",
-        xll: "1400px",
+        xl: "1280px",      
         "2xl": "1536px",
       },
     },